WIP: multidimensional performance tests
